--- src/components/LandingPage/Features/index.module.css
@@ -1,217 +1,0 @@
-.container {
-    width: Min(1250px, 100%);
-    margin: auto;
-    position: relative;
-    padding: 150px 0 200px 0;
-    --swiper-theme-color: var(--ifm-color-primary);
-    z-index: 1;
-}
-.anchor {
-    display: block;
-    position: absolute;
-    margin-top: -120px;
-    padding-bottom: 120px;
-    visibility: hidden;
-}
-.title {
-    width: Min(1250px, 90%);
-    margin: auto;
-    font-weight: 700;
-    font-size: 40px;
-    line-height: 50px;
-    color: black;
-}
-
-.cards {
-    width: Min(1250px, 90%);;
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    grid-auto-rows: 1fr;
-    grid-column-gap: 4px;
-    grid-row-gap: 4px;
-    margin: 3rem auto;
-}
-
-.card {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    padding: 0 48px;
-    width: 100%;
-    margin: auto 0;
-    height: 200px;
-    background: rgba(255, 255, 255, 0.5);
-    border: 1px solid #FFFFFF;
-    box-sizing: border-box;
-    backdrop-filter: blur(5px);
-    border-radius: 12px;
-    transition: all var(--ifm-transition-custom-speed) var(--ifm-transition-custom-curve);
-}
-
-.card:hover {
-    padding-bottom: 32px;
-    background: var(--ifm-color-primary);
-    border: 1px solid var(--ifm-color-primary);
-}
-
-.card:hover,
-.card:hover .cardTitle,
-.card:hover .cardBody,
-.card:hover .informationIcon {
-    color: white;
-    text-decoration: none;
-    transition: all var(--ifm-transition-custom-speed) var(--ifm-transition-custom-curve);
-}
-
-.cardContainer {
-    width: 90%;
-}
-
-.cardTitle {
-    font-weight: 700;
-    font-size: 24px;
-    line-height: 32px;
-    color: var(--ifm-color-primary);
-    margin-bottom: 0.7rem;
-}
-
-.cardBody {
-    font-weight: 400;
-    font-size: 14px;
-    line-height: 22px;
-    color: black;
-    margin-bottom: 0;
-}
-
-.actionButtonContainer, .actionButtonContainerMobile {
-    width: Min(1250px, 90%);
-    margin: auto;
-    justify-content: flex-start;
-}
-
-.actionButtonContainer {
-    display: flex;
-}
-
-.actionButtonContainerMobile {
-    display: none;
-}
-
-.actionButton {
-    margin: 0;
-    text-align: center;
-    padding: 16px 24px;
-    font-weight: 700;
-    font-size: 16px;
-    line-height: 22px;
-    letter-spacing: 1px;
-    border: 2px solid black;
-    border-radius: 10px;
-    color: black;
-    text-decoration: none;
-    transition: all var(--ifm-transition-custom-speed) var(--ifm-transition-custom-curve);
-}
-
-.actionButton:hover {
-    border: 2px solid #3B00B9;
-    background-color: #3B00B9;
-    color: white;
-    text-decoration: none;
-}
-
-.informationIcon {
-    color: var(--ifm-color-primary);
-    margin: auto;
-    width: 24px;
-    height: 24px;
-}
-
-.scrollContainer {
-    display: none;
-    width: 100vw;
-    margin: 32px 0 12px 0;
-}
-
-.mobileCardsContainer {
-    position: relative;
-    z-index: 1000;
-    display: flex;
-    overflow-x: scroll;
-    scroll-padding: 50% 0;
-    scroll-snap-type: x mandatory;
-    scrollbar-width: thin;
-}
-
-.mobileCardsContainer::-webkit-scrollbar {
-    width: 12px;
-}
-
-.mobileCardsContainer::-webkit-scrollbar-track {
-    background: rgba(24, 24, 24, 0%);
-    margin: 0 calc((100vw - Min(1100px, 90vw)) / 2);
-    border-radius: 20px;
-}
-
-.mobileCardsContainer::-webkit-scrollbar-thumb {
-    background-color: rgba(24, 24, 24, 20%);
-    border-radius: 20px;
-    border: 4px solid var(--ifm-background-color);
-}
-
-.cardWrapper .card {
-    width: 300px;
-    padding: 0 32px;
-}
-
-.cardWrapper {
-    scroll-snap-align: none;
-    display: inline-block;
-    margin: 0 4px 10px 0;
-}
-
-.cardWrapper:first-child {
-    padding-left: calc((100vw - Min(1100px, 90vw)) / 2);
-}
-
-.cardWrapper:last-child {
-    margin-right: 0;
-    padding-right: calc((100vw - Min(1100px, 90vw)) / 2);
-}
-
-@media screen and (max-width: 996px) {
-    .container {
-        padding: 48px 0 80px 0;
-    }
-
-    .cards, .actionButtonContainer {
-        display: none;
-    }
-
-    .actionButtonContainerMobile {
-        display: flex;
-    }
-
-    .scrollContainer {
-        display: block;
-    }
-
-    .title {
-        font-size: 32px;
-        line-height: 42px;
-    }
-
-    .informationIcon {
-        width: 16px;
-        height: 16px;
-    }
-
-    .actionButton {
-        font-size: 14px;
-        line-height: 22px;
-        padding: 14px 20px;
-    }
-    .anchor {
-        margin-top: -100px;
-        padding-bottom: 100px;
-    }
-}
