--- src/components/LandingPage/Showcase/index.module.css
@@ -1,333 +1,0 @@
-.container {
-    background-color: transparent;
-    padding: 0;
-    position: relative;
-    --swiper-theme-color: white;
-    --swiper-pagination-bullet-inactive-color: white;
-    --swiper-pagination-bullet-inactive-opacity: 0.30;
-    z-index: 2;
-}
-.anchor {
-    display: block;
-    position: absolute;
-    margin-top: -120px;
-    padding-bottom: 120px;
-    visibility: hidden;
-}
-.showcaseContainer {
-    width: 100vw;
-    padding: 50px 0;
-    margin: auto;
-    position: relative;
-}
-
-.header {
-    display: flex;
-    justify-content: flex-start;
-    align-items: flex-start;
-    flex-direction: column;
-    width: Min(1100px, 90%);
-    margin: auto auto 50px auto;
-}
-
-.title {
-    font-weight: 700;
-    font-size: 60px;
-    line-height: 70px;
-    color: white;
-    z-index: 1000;
-    position: relative;
-    margin-bottom: 20px;
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-}
-
-.textCycling {
-    position: absolute;
-    margin-left: 15px;
-}
-
-.headerBody {
-    position: relative;
-    z-index: 1000;
-}
-
-.body {
-    color: white;
-    font-weight: 400;
-    font-size: 18px;
-    line-height: 28px;
-    width: 434px;
-    z-index: 1000;
-}
-
-.callToAction {
-    display: flex;
-    flex-direction: row;
-    color: white;
-    font-weight: 700;
-    font-size: 18px;
-    line-height: 26px;
-    z-index: 1000;
-    transition: all var(--ifm-transition-custom-speed) var(--ifm-transition-custom-curve);
-}
-
-.callToAction:hover {
-    color: rgba(182, 182, 182, 1);
-    text-decoration: none;
-}
-
-.callToAction p {
-    margin-bottom: 0;
-}
-
-.callToAction svg {
-    min-width: 24px;
-    height: 24px;
-    display: inline-flex;
-    align-self: center;
-    margin: 0 8px auto 0;
-}
-
-
-.actionButtonContainer {
-    width: Min(1100px, 90%);
-    margin: 3rem auto;
-}
-
-.actionButton {
-    text-align: center;
-    display: inline-flex;
-    padding: 16px 24px;
-    font-weight: bold;
-    font-size: 16px;
-    line-height: 22px;
-    border: 2px solid white;
-    border-radius: 10px;
-    color: white;
-    text-decoration: none;
-    position: relative;
-    z-index: 1000;
-    transition: all var(--ifm-transition-custom-speed) var(--ifm-transition-custom-curve);
-}
-
-.actionButton:hover {
-    border: 2px solid white;
-    background-color: white;
-    color: #3B00B9;
-    text-decoration: none;
-}
-
-.BGShape {
-    position: fixed;
-    width: 100vw;
-    top: 0;
-    left: 0;
-    z-index: 1000;
-}
-
-.cardContainer {
-    position: relative;
-    z-index: 1000;
-    display: flex;
-    overflow-x: scroll;
-    scroll-padding: 50% 0;
-    scroll-snap-type: x mandatory;
-    scrollbar-width: thin;
-    scrollbar-color: white #3b00b9;
-}
-
-.cardContainer::-webkit-scrollbar {
-    width: 12px;
-}
-
-.cardContainer::-webkit-scrollbar-track {
-    background: #3b00b9;
-    margin: 0 calc((100vw - Min(1100px, 90vw)) / 2);
-    border-radius: 20px;
-}
-
-.cardContainer:hover::-webkit-scrollbar-thumb {
-    background-color: rgba(24, 24, 24, 60%);
-}
-
-.cardContainer::-webkit-scrollbar-thumb {
-    background-color: #3b00b9;
-    border-radius: 20px;
-    border: 3px solid #3B00B9;
-}
-
-.cardWrapper {
-    scroll-snap-align: none;
-    display: inline-block;
-    margin: 0 20px 20px 0;
-}
-
-.cardWrapper:first-child {
-    padding-left: calc((100vw - Min(1100px, 90vw)) / 2);
-}
-
-.cardWrapper:last-child {
-    margin-right: 0;
-    padding-right: calc((100vw - Min(1100px, 90vw)) / 2);
-}
-
-.dappContainer {
-    background: white;
-    display: inline-block;
-    height: 100%;
-    width: 440px;
-    border-radius: 12px;
-}
-
-.dappContainer:hover {
-    text-decoration: none;
-}
-
-.dappHeader {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: flex-start;
-    height: 100px;
-    margin: 24px 24px 16px;
-}
-
-.dappIcon {
-    display: flex;
-    justify-content: center;
-    align-items: flex-start;
-    margin-right: 16px;
-}
-
-.dappInformation {
-    height: 100%;
-}
-
-.dappName {
-    font-weight: 500;
-    font-size: 24px;
-    line-height: 32px;
-}
-
-.dappStats {
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-    letter-spacing: -0.15px;
-    margin: 4px 0 8px 0;
-    color: #888888;
-}
-
-.dappWebsite {
-    display: inline-block;
-    font-weight: 400;
-    font-size: 12px;
-    line-height: 16px;
-    letter-spacing: -0.15px;
-    color: #262626;
-    padding: 2px 6px;
-    background: #F2F2F2;
-    border-radius: 5px;
-    transition: color var(--ifm-transition-custom-speed);
-}
-
-.dappWebsite:hover {
-    color: #c2c6c7;
-    text-decoration: none;
-}
-
-.dappMedia {
-    display: flex;
-    align-items: flex-end;
-    width: 100%;
-    height: 218px;
-    padding: 5% 10% 5% 10%;
-}
-
-.dappVideo, .dappScreenshot {
-    object-fit: contain;
-    object-position: center;
-    width: 100%;
-    height: 100%;
-}
-
-.dappBody {
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-    letter-spacing: -0.15px;
-    color: #888888;
-    margin: 24px;
-}
-
-
-@media screen and (max-width: 996px) {
-
-    .container {
-        background-color: #3b00b9;
-    }
-
-    .BGShape {
-        display: none;
-    }
-
-    .showcaseContainer {
-        margin: auto;
-        padding: 100px 0;
-    }
-
-    .background {
-        display: none;
-    }
-
-    .title, .body {
-        width: Min(434px, 90%);
-    }
-
-    .title {
-        font-size: 40px;
-        line-height: 50px;
-        flex-direction: column;
-        margin-bottom: 70px;
-    }
-
-    .textCycling {
-        margin: 0;
-        width: 90vw;
-    }
-
-    .actionButtonContainer {
-        margin: 3rem auto;
-    }
-
-    .actionButton {
-        padding: 14px 24px;
-        font-size: 14px;
-        line-height: 22px;
-    }
-
-    .cardWrapper {
-        scroll-snap-align: center;
-        margin: 0 10px 10px 0;
-    }
-
-    .dappContainer {
-        width: 330px;
-    }
-
-    .dappName {
-        font-size: 20px;
-        line-height: 28px;
-    }
-
-    .dappStats {
-        font-size: 16px;
-        line-height: 20px;
-    }
-    .anchor {
-        margin-top: -100px;
-        padding-bottom: 100px;
-    }
-}
